Initial commit for positioners, included an Identity positioner and a Stacked positioner. Because positioners may add attributes to data objects that affect the calculation of scale domains, the logic for calculating the domain has been moved to the positioner objects.
Note this is a first pass and I'd appreciate feedback on the general
implementation. Of particular note is the fact that only the line geom
currently supports stacking and other geoms will require some effort to be made
compatible. Also note the specialized scale method now present in the Area and
Line geoms suggesting a more generalized version may be more appropriate.